+++ tests/testthat/test-formatter.r
@@ -227,6 +227,14 @@ test_that("Single NA gives NA", {
 })
 
 
+test_that("default accuracy works for range of inputs", {
+  x <- c(0.1, 0.2, 0.5)
+  expect_equal(percent(x / 100), c("0.1%", "0.2%", "0.5%"))
+  expect_equal(percent(x / 10), c("1%", "2%", "5%"))
+  expect_equal(percent(x), c("10%", "20%", "50%"))
+  expect_equal(percent(x * 10), c("100%", "200%", "500%"))
+})
+
 # Dollar formatter --------------------------------------------------------
 
 test_that("negative comes before prefix", {